pyproject.toml
[project]
name = "walmart-genai-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Ruijiao Guo", email = "ruijiao.guo@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.1",
    "openai>=2.17.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]

[project.scripts]
walmart-genai-system = "walmart_genai_system:main"

[build-system]
requires = ["uv_build>=0.9.30,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "black>=26.1.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.15.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
fix = true

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]